fix: resolve circular dependency in AI providers causing build failure
Issue #231: The production build was failing with 'Cannot access hy before
initialization' when prerendering the /settings page. This was caused by
a circular dependency in the AI provider modules:

- src/ai/providers/index.ts exported from ./claude, ./openai, ./zaic
- Each of those files imported AIProviderConfig and DEFAULT_MODELS from index.ts

Fixed by extracting shared types to a separate types.ts file and updating
all provider files to import from the new file instead of the index.
