+++ src/ai/providers/index.ts
@@ -9,67 +9,17 @@
  * The AI functionality should only be used server-side or with proper fallback handling.
  */
 
+// Import types from separate file (to avoid circular dependencies)
+import { DEFAULT_CONFIGS, DEFAULT_MODELS, type AIProviderConfig, type AIProvider } from './types';
+
+// Re-export types
+export { DEFAULT_CONFIGS, DEFAULT_MODELS };
+export type { AIProviderConfig, AIProvider };
+
 // Re-export Claude, OpenAI and Z.ai providers
 export * from './claude';
 export * from './openai';
 export * from './zaic';
-
-/**
- * Supported AI providers
- */
-export type AIProvider = 'google' | 'openai' | 'anthropic' | 'zaic' | 'custom';
-
-/**
- * Configuration options for AI providers
- */
-export interface AIProviderConfig {
-  provider: AIProvider;
-  model?: string;
-  apiKey?: string;
-  temperature?: number;
-  maxOutputTokens?: number;
-}
-
-/**
- * Default model configurations
- */
-export const DEFAULT_MODELS: Record<string, string> = {
-  google: 'gemini-1.5-flash-latest',
-  openai: 'gpt-4o-mini',
-  anthropic: 'claude-3-haiku-20240307',
-  zaic: 'default',
-};
-
-/**
- * Default configurations for each provider
- */
-export const DEFAULT_CONFIGS: Record<AIProvider, Partial<AIProviderConfig>> = {
-  google: {
-    model: DEFAULT_MODELS.google,
-    temperature: 0.7,
-    maxOutputTokens: 8192,
-  },
-  openai: {
-    model: DEFAULT_MODELS.openai,
-    temperature: 0.7,
-    maxOutputTokens: 8192,
-  },
-  anthropic: {
-    model: DEFAULT_MODELS.anthropic,
-    temperature: 0.7,
-    maxOutputTokens: 8192,
-  },
-  zaic: {
-    model: DEFAULT_MODELS.zaic,
-    temperature: 0.7,
-    maxOutputTokens: 8192,
-  },
-  custom: {
-    model: DEFAULT_MODELS.google,
-    temperature: 0.7,
-    maxOutputTokens: 8192,
-  },
-};
 
 /**
  * Current active provider configuration
